requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-core>=0.1.10
openai>=1.6.0
python-dotenv>=1.0.0
chromadb>=0.4.22
tiktoken>=0.5.2
pydantic>=2.5.2